qv-larga-cetagua-api/requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
Django==5.0.2
django-cors-headers==4.3.1
django-simple-history==3.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
et-xmlfile==1.1.0
Faker==23.2.1
fiona==1.9.5
geopandas==0.14.3
idna==3.6
numpy==1.26.4
openpyxl==3.1.2
packaging==23.2
pandas==2.2.0
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pyproj==3.6.1
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
shapely==2.0.3
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.0
uritemplate==4.1.1
python-dotenv==1.0.1
statsmodels==0.14.1
scikit-learn==1.4.1.post1
joblib==1.3.2